Created firmware type; many bug fixes

+++ SQL/awakemud.sql
@@ -411,7 +411,7 @@ CREATE TABLE `matrix_files` (
   `last_decay_time` mediumint(5) unsigned default '0',
 
   `creator_idnum` mediumint(5) unsigned default '0',
-  `in_obj_vnum` mediumint(5) unsigned default '0',
+  `in_obj_vnum` bigint(5) unsigned default '0',
   KEY(`idnum`)
 );
